Add flake.nix so we can nixify the python setup
This doesn't break any existing workflows.
I use NixOS so it's technically not possible to install
the packages mentioned in the requirments.txt.
The only way for me to develop is by fetching these
dependencies using nix.
The dependencies are manually copied from requirements.
In follow-up PRs, I will make sure that the dependencies
are in sync with the requirements.txt