Fix: Update Sphinx Dependency Names (#159)
This commit resolves an issue #158 where the dependency names for sphinx-panels and sphinx-rtd-dark-mode in the requirements file did not match those used on some platforms, potentially causing confusion or errors when attempting to install the dependencies. The requirements file has been updated to include both versions of each dependency's name to ensure cross-compatibility.