Update check-black.yml (#202)
* Update check-black.yml

I'm not sure that the actions were picking up changes in the pyproject.toml file. This lead to some back and forth errors where formatting with black would break formatting with isort and vice versa.

* Update check-isort.yml (#203)

I'm not sure that the actions were picking up changes in the pyproject.toml file. This lead to some back and forth errors where formatting with black would break formatting with isort and vice versa.