Dockerize the Automata Project (resolves #285)
This commit introduces a Dockerfile that allows the project to be run in a Docker container. The Dockerfile includes instructions for installing all necessary dependencies and running the application.

Additionally, instructions on how to build and run the Docker image have been added as comments at the top of the Dockerfile for user's convenience.

This change facilitates development and deployment processes, and ensures consistency across different environments.