Dockerize the Automata Project (resolves #285) (#286)
* Dockerize the Automata Project (resolves #285)

This commit introduces a Dockerfile that allows the project to be run in a Docker container. The Dockerfile includes instructions for installing all necessary dependencies and running the application.

Additionally, instructions on how to build and run the Docker image have been added as comments at the top of the Dockerfile for user's convenience.

This change facilitates development and deployment processes, and ensures consistency across different environments.

* fix: maintain pip installation of Poetry due to Docker issues

The requested changes for Poetry were attempted, including both the installation via curl and applying the 'poetry run black .' command. Despite extensive efforts, these changes were not successful.

This does not reflect on the quality of the proposed ideas but rather the unpredictable behavior of Docker in certain contexts. Encountered significant Docker issues that prevented further progress.

Decided to commit the current working state that employs pip for Poetry installation. Future adjustments may be needed when Docker issues are resolved.
Resolves #285

* Remove: Delete the OLD_Dockerfile.

Resolves #285